fix(dbradham): address various issues and clean code before release
