+++ apps/dev-recruiters/src/components/modules/OpportunitiesAggregatorWithRoles/OpportunitiesAggregator.tsx
@@ -4,13 +4,15 @@ import {
 
 import * as React from "react";
 import Slider from "react-slick";
-import "slick-carousel/slick/slick-theme.css";
+// uncomment the below line at your own risk, it has broken /join page
+//import "slick-carousel/slick/slick-theme.css";
 import "slick-carousel/slick/slick.css";
 import BoxContainer from "../../common/BoxContainer";
 import FilteringComponent, {
   FilteringComponentProps
 } from "./filtering/FilteringComponent";
 import ShortCard from "./ShortCard";
+import '../hide-slick-dots.css';
 interface Props extends FilteringComponentProps {}
 
 const OpportunitiesAggregator: React.FunctionComponent<Props> = ({
