+++ apps/dev-recruiters/src/components/modules/hide-slick-dots.css
@@ -1,0 +1,3 @@
+.slick-dots li {
+    display: none;
+}
