--- apps/ideaspace/src/components/modules/IdeaGeneratorButton/AppIdeaGeneratorButton/StyledAppIdeaGeneratorButton.js
@@ -1,34 +1,0 @@
-import styled from 'styled-components';
-
-export const ButtonBackground = styled.div`
-  height: 100%;
-  min-height: 300px;
-  max-height: 300px;
-  display: inline;
-  background-size: cover;
-  border-radius: 30px;
-  align-content: center;
-  overflow: initial;
-`;
-
-export const ButtonTitleDiv = styled.div`
-  color: white;
-  width: 100%;
-  border: solid;
-  border-top: 0;
-  border-left: 0;
-  border-right: 0;
-`;
-
-export const ButtonText = styled.p`
-  color: white;
-`;
-
-export const ButtonDescriptionDiv = styled.div`
-  padding-top: 9%;
-`;
-
-export const ButtonContinueDiv = styled.div`
-  text-align: right;
-  padding-top: 1%;
-`;
