--- apps/ideaspace/src/components/modules/IdeaGeneratorButton/AppIdeaGeneratorButton/index.js
@@ -1,1 +1,0 @@
-export { default } from "./AppIdeaGeneratorButton";
