+++ apps/ideaspace/src/components/modules/IdeaspaceNavCard/IdeaspaceNavCard.js
@@ -18,8 +18,10 @@ const IdeaspaceNavCard = ({
         background: `linear-gradient(${cardBackgroundColor}, url(${cardBackgroundImage})`,
       }}
     >
+      <div>
       <CardTitleDiv>{cardTitle}</CardTitleDiv>
       <CardText>{cardDescription}</CardText>
+      </div>
       <CardContinueDiv>
         <CardText>CONTINUE &#8594;</CardText>
       </CardContinueDiv>
