--- apps/ideaspace/src/components/modules/Nav/nav.css
@@ -1,26 +1,0 @@
-.nav {
-  background-color: #1c1c1c;
-  color: white;
-  height: 50px;
-  display: flex;
-  justify-content: left;
-  align-items: center;
-  padding-left: 2rem;
-}
-
-.nav a {
-  text-decoration: none;
-  color: white;
-  margin-left: 10px;
-}
-
-@media (orientation: portrait) {
-  .nav {
-    font-size:1rem;
-  }
-  .nav a {
-    padding-left: .5rem;
-  }
-}
-
-
