+++ apps/ideaspace/src/components/modules/WelcomePage/StyledWelcomePage.js
@@ -1,13 +1,10 @@
 import styled from 'styled-components';
 
-export const ButtonArea = styled.div`
-  width: 100%;
-  min-height: 50vh;
-  height: 1px;
+export const ButtonArea = styled.aside`
   max-width: 1700px;
   display: flex;
   flex-direction: row;
-  justify-content: space-around;
+  justify-content: space-between;
   align-items: flex-start;
   margin-left: auto;
   margin-right: auto;
@@ -19,14 +16,10 @@ export const ButtonArea = styled.div`
     width: 80%;
     min-height: 60vh;
   }
-`;
 
-export const StyledLink = styled.a`
-  height: 100%;
-  color: 'white';
-  @media (orientation: portrait) {
-    width: 100%;
-    height: 40%;
+  @media (max-width: 575px) {
+    flex-direction: column;
+    justify-content: space-between;
   }
 `;
 
@@ -80,31 +73,6 @@ export const Header = styled.div`
 export const BackgroundImage = styled.img`
   width: 100%;
   height: 50%;
-`;
-
-export const ButtonWrapper = styled.div`
-  width: 23%;
-  display: inline-grid;
-  flex-direction: row;
-  justify-content: space-between;
-  align-items: center;
-  /* background: #C4C4C4; */
-  /* font-weight: bold; */
-  color: ${({ theme }) => theme.colors.NEUTRAL_2};
-  border: 0 none;
-  border-radius: 30px;
-  cursor: pointer;
-  padding: 2rem;
-  margin: 10px 5px;
-  font-size: 2rem;
-  font-family: 'Abel';
-  background-color: ${({ theme }) => theme.colors.NEUTRAL_1};
-  font-color: 'white';
-  @media (orientation: portrait) {
-    font-size: 1.5rem;
-    flex-direction: column;
-    opacity: 0.2;
-  }
 `;
 
 export const WelcomeMessage = styled.div`
